[#43] move get_secret to filters
